+++ languageServer/package.json
@@ -6,17 +6,9 @@
     "node": "*"
   },
   "dependencies": {
-    "@swc/core": "1.2.247",
-    "@swc/jest": "0.2.22",
-    "@types/jest": "28.1.8",
-    "fs-jetpack": "4.3.1",
-    "jest": "29.0.1",
-    "jest-environment-node": "29.0.2",
     "jsonc-parser": "3.1.0",
     "jsonpath-plus": "7.0.0",
-    "lint-staged": "13.0.3",
     "lodash": "4.17.21",
-    "ts-jest": "28.0.8",
     "vscode-languageserver": "8.0.2",
     "vscode-languageserver-textdocument": "1.0.4",
     "vscode-uri": "3.0.3",
@@ -32,5 +24,13 @@
     "dev": "tsc -b --watch",
     "test": "jest --collect-coverage"
   },
-  "devDependencies": {}
+  "devDependencies": {
+    "@swc/core": "1.2.247",
+    "@swc/jest": "0.2.22",
+    "@types/jest": "28.1.8",
+    "fs-jetpack": "4.3.1",
+    "lint-staged": "13.0.3",
+    "jest": "29.0.1",
+    "jest-environment-node": "29.0.2"
+  }
 }
